+++ biopima/src/app/api/sensor-readings/route.ts
@@ -4,7 +4,7 @@ const baseUrl = process.env.BASE_URL;
 export async function GET() {
   try {
     
-    const response = await fetch(`${baseUrl}/sensor-readings/`);
+    const response = await fetch(`${baseUrl}/api/sensor-readings/`);
     const result = await response.json();
     return new Response(JSON.stringify(result), {
       status: 200,
