+++ biopima/src/app/login/page.test.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { render, screen, fireEvent, waitFor } from "@testing-library/react";
 import "@testing-library/jest-dom";
-import SignInPage from "./page";
+import SignInPage from "./login";
 import { useRouter, useSearchParams } from "next/navigation";
 import { useLogin } from "../hooks/useFetchLogin";
